// server/src/api/repositories/chat/enum/chat.enum.ts
export enum ChatEnum {
  USER_ALREADY_IN_CHAT= 'User already in chat',
  NO_USER_FOUNDED = 'No user founded',
  NO_CHAT_NAME = 'Provide chat name',
  NO_CHAT_ID = 'Provide chat id',
  NO_USER_ID = 'Provide user id',
  CHAT_SUCCESSFULLY_CREATED = 'Chat successfully created',
  MESSAGE_SUCCESSFULLY_SENT = 'Message successfully sent',
  MESSAGE_SUCCESSFULLY_READ = 'Message successfully read',
  SOME_FIELDS_IS_NULL = 'Some fields is null',
  NO_CHAT_FOUND = 'No chat found by this id',
  USER_SUCCESSFULLY_ADDED_TO_CHAT = 'User successfully added to chat room'
}